requirements.txt
Flask==2.3.2
pymongo==4.10.1
python-dotenv==1.0.0
requests==2.31.0
passlib[bcrypt]==1.7.4
pyjwt==2.6.0
marshmallow==3.20.1
